Added initial protobuf definition.
Protobuf will be used for both intermediate files during processing and the final dataset output. The final dataset output will consist of a series of CitationGroup messages. Each CitationGroup message contains a number (need to figure out how many we can fit given the maximum sizes of a message) of citations. Because processing will try to find when a citation was first discovered and when it is removed this should hopefully deduplicate a lot of stuff for us already.

Signed-off-by: Zoe Nickson <mnickson@sidingsmedia.com>