New repo structure (#5)
* Initialised repository with new structure

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Add protobuf definitions using vcpkg

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Added initial parser implementation.

This implementation is not as fully featured as the original in WikiOpenCite/parser, it only extracts the title currently. A future PR will more fully implement the parser to include other relevant information (i.e. doi, pmid...).

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Fixed coverage CI build

Changed from Coverage build type to Release build type as this was causing issues with protobuf

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Actually fix the coverage issues

Setting the coverage mode to release wasn't the fix. Instead we just remove `-fkeep-inline-functions` to prevent linking errors with protobuf.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Removed windows test

We don't really intend on supporting windows, the build was slow and failed anyway.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

---------

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>