Completed parsing of citation templates (#9)
* Added parsing of identifiers

This commit introduces parsing of identifiers. To support greater flexibility of use, optional configuration may now be passed to the parser. Currently this controls if an exception should be thrown when an invalid identifier is encountered.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Added additional test to cover integer too large

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Added URL extraction to parser.

Currently this is limited to url and archive-url as a result of the protobuf message however we can always extend this in future.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

---------

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>