Initial implementation of the extractor
The extractor with citation deduplication has been kinda implemented. It has one main limitation. If a citation is added in revision A and then removed in revision B, readded in revision C and removed again in D it will be marked as being added in A and removed in D. It is therefore a more of a "last seen" rather than a removed in.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>