Initial Extractor Implementation (#15)
* Add initial extractor structure

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Bumped proto to 0.3.0

The proto update brought in the new page datatype so the rtype of the extractors have been updated. Also changed to a map for the extracted citations for some improved performance later when deduplicating.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Update reuse to include test data

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Add prefix to test cases for easier ID

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Fix linking errors

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Placeholder parsing methods

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Initial extractor test cases

TODO: Check timestamp and other fields for revisions.
Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Remove additional XML data

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Initial implementation of the extractor

The extractor with citation deduplication has been kinda implemented. It has one main limitation. If a citation is added in revision A and then removed in revision B, readded in revision C and removed again in D it will be marked as being added in A and removed in D. It is therefore a more of a "last seen" rather than a removed in.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Modified test case to handle multiple revisions with no change

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Added BZ2 test case

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Implemented error handling

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Make sure only revisions that are referenced appear in the list (mostly)

This does have a slight catch for when we handle citations that are removed and then readded but this should't be too big a deal

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Added docs

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

* Make sure we don't include revisions that arent referenced

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>

---------

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>