Refactored DumpParser for clarity and inheritance
DumpParser has been refactored to make it more readable (it wasn't exactly pretty before). Also because the stream parser will need to inherit from it, it needed to be made more suitable for inheritance.

To inherit from DumpParser, the inheriting class should override the Store method to provide it's own way of storing the produced page and revisions. Once Store is called for a given page, the DumpParser has finished processing them so the inheriting class is free to do as it wishes with it.

Signed-off-by: Zoe Nickson <mjn6@st-andrews.ac.uk>